Various changes since last git push
fabSuiteCom put within scope
. check 3rd party FS status
. insert i18n
. progress warning
. control form buttons
. fs login use prefs
jobsts 
. pull 3rd party stations
. pull forms for permissions
. check fs on status input
. mobile code return to input field
. additional error checking
. set user permissions on base forms, skip instances on name collection
. rfMobile edits for single input area changes
. slim kiss import work
. kiss import progress warnings
kiss import
. set up import table filters as per STS Xe
kiss module
. filtering as early as possible in import
scoping changes throughout
additional i18n messages
secGroupBase permissions bug where selection removed unnecessarily
set up base code for default groups
app_permiss_detail save failing wrong index start 
app_permissions set up no edit status for keys on default corporate,
office, shop logins
app_user_group_list protect deletion default groups
kiss_import dialog
. add filter fields
. button access when accessing fields 
. lay groundwork for splitting for use by web client
kiss_option_import
. move table to base of view for larger table
. set import actions for each incoming mark
load_critera progress warnings
STS globals set filters for corp/office/shop

+++ servoySecurity/forms/secGroupBase.js
@@ -40,6 +40,8 @@ function onDataChangeGroupKeys(oldIDs, newIDs, event) {
 	newIDs = (newIDs) ? new String(newIDs).split('\n') : [];				//	split the new id list into an array by carriage return
 	newIDs.sort();															//	sort the array
 	
+	if (newIDs.length == oldIDs.length){return}
+	
 	if(newIDs.length > oldIDs.length){										//	check difference: id was ADDED
 		id = newIDs[newIDs.length - 1];									//	determine the id, starting with the last element in the array
 		for(i in oldIDs){													//	iterate by the old ids
@@ -54,22 +56,20 @@ function onDataChangeGroupKeys(oldIDs, newIDs, event) {
 			databaseManager.saveData(fs.getSelectedRecord());				//	save data
 		}
 	} else {																//	id was REMOVED	
-		//id = oldIDs[oldIDs.length - 1];									//	determine the id, starting with the last element in the array
-		/**for(i in newIDs){													//	iterate by the new ids
-			if(oldIDs[i] != newIDs[i])										//	until the corresponding old id doesn't match
+		id = oldIDs[oldIDs.length - 1];									//	determine the id, starting with the last element in the array
+		for(i in newIDs){													//	iterate by the new ids
+			if(oldIDs[i] != newIDs[i]){										//	until the corresponding old id doesn't match
 				id = oldIDs[i];												//	which gives the id that was removed
-		}*/
-		id = oldIDs.pop();
-		while (newIDs.indexOf(id) != -1){
-			id = oldIDs.pop();
+				null;
+				break;
+			}
 		}
-		if(fs.find()){														//	search the foundset
+		if(id && fs.find()){														//	search the foundset
 			fs[idColumnName] = id;											//	by the fk column
 			if(fs.search())																					
 				fs.deleteRecord();											//	delete record
 		}
 	}
-	//updateUI();																//	update user interface
 	updateUI(event);																//	update user interface
 	return true;															//	allow data change
 }
